+++ hutool-http/src/main/java/com/xiaoleilu/hutool/http/HttpResponse.java
@@ -13,6 +13,7 @@ import java.util.zip.GZIPInputStream;
 
 import com.xiaoleilu.hutool.convert.Convert;
 import com.xiaoleilu.hutool.io.FastByteArrayOutputStream;
+import com.xiaoleilu.hutool.io.IORuntimeException;
 import com.xiaoleilu.hutool.io.IoUtil;
 import com.xiaoleilu.hutool.util.StrUtil;
 
@@ -202,22 +203,30 @@ public class HttpResponse extends HttpBase<HttpResponse> {
 	}
 	
 	/**
-	 * 读取主体
+	 * 读取主体，忽略EOFException异常
 	 * @param in 输入流
 	 * @return 自身
-	 * @throws IOException
+	 * @throws IORuntimeException IO异常
 	 */
-	private void readBody(InputStream in) throws IOException{
+	private void readBody(InputStream in) throws IORuntimeException{
 		if(isGzip() && false == (in instanceof GZIPInputStream)){
-			in = new GZIPInputStream(in);
+			try {
+				in = new GZIPInputStream(in);
+			} catch (IOException e) {
+				throw new IORuntimeException(e);
+			}
 		}
 		
 		int contentLength  = Convert.toInt(header(Header.CONTENT_LENGTH), 0);
 		this.out = contentLength > 0 ? new FastByteArrayOutputStream(contentLength) : new FastByteArrayOutputStream();
 		try {
 			IoUtil.copy(in, this.out);
-		} catch (EOFException e) {
-			//忽略读取HTTP流中的EOF错误
+		} catch (IORuntimeException e) {
+			if(e.getCause() instanceof EOFException) {
+				//忽略读取HTTP流中的EOF错误
+			}else {
+				throw e;
+			}
 		}
 	}
 	
@@ -237,8 +246,8 @@ public class HttpResponse extends HttpBase<HttpResponse> {
 		//非同步状态转为同步状态
 		try {
 			this.readBody(this.in);
-		} catch (IOException e) {
-			if(e instanceof FileNotFoundException){
+		} catch (IORuntimeException e) {
+			if(e.getCause() instanceof FileNotFoundException){
 				//服务器无返回内容，忽略之
 			}else{
 				throw new HttpException(e);